chore: fix cops with the wrong namespace
As part of the PR for adding support for Ruby 3.3 (#205) we bumped the
`rubocop-rspec` dependency to `~> 2.26`, which depends on
`rubocop-capybara` `~> 2.17` and `rubocop-factory_bot` `~> 2.22`. On
those versions, the `RSpec` namespace was removed from their cops hence
a warning was being logged when running `rubocop` with the last version
of this gem.

Related PRs:
- https://github.com/rubocop/rubocop-factory_bot/commit/812771eb33453e40f51c7a1cc993d7c8b6f5b33a
- https://github.com/rubocop/rubocop-capybara/commit/e36f30457b6a49d1e04847d2337186dc0d03a435